+++ src/main/java/org/accounting/system/repositories/metric/MetricRepository.java
@@ -168,7 +168,7 @@ public class MetricRepository extends AccessibleModulator<Metric, ObjectId> {
         return projectionQuery;
     }
 
-    public PanacheQuery<MetricProjection> findByProjectIdAndGroupId(final String externalId, final String groupId, int page, int size) {
+    public PanacheQuery<MetricProjection> findByProjectIdAndGroupId(final String externalId, final String groupId, int page, int size, String start, String end) {
 
         var filters = Array.of(Filters.regex("resource_id", "\\b" + externalId + "\\b"+"(?![-])"));
 
@@ -177,6 +177,11 @@ public class MetricRepository extends AccessibleModulator<Metric, ObjectId> {
         Bson lookup = Aggregates.lookup("MetricDefinition", "metric_definition_id", "_id", "metric_definition");
 
         filters = filters.append(Filters.eq("group_id", groupId));
+
+        if(ObjectUtils.allNotNull(start, end) && Utility.isDate(start, end) && Utility.isBefore(start, end)) {
+            filters = filters.append(Filters.and(Filters.gte("time_period_start", Utility.stringToInstant(start)), Filters.lte("time_period_start", Utility.stringToInstant(end))));
+            filters = filters.append(Filters.and(Filters.gte("time_period_end", Utility.stringToInstant(start)), Filters.lte("time_period_end", Utility.stringToInstant(end))));
+        }
 
         Bson regex = Aggregates.match(Filters.and(filters));
 
@@ -199,7 +204,7 @@ public class MetricRepository extends AccessibleModulator<Metric, ObjectId> {
         return projectionQuery;
     }
 
-    public PanacheQuery<MetricProjection> findByProjectIdAndUserId(final String externalId, final String userId, int page, int size) {
+    public PanacheQuery<MetricProjection> findByProjectIdAndUserId(final String externalId, final String userId, int page, int size, String start, String end) {
 
         var filters = Array.of(Filters.regex("resource_id", "\\b" + externalId + "\\b"+"(?![-])"));
 
@@ -208,6 +213,11 @@ public class MetricRepository extends AccessibleModulator<Metric, ObjectId> {
         Bson lookup = Aggregates.lookup("MetricDefinition", "metric_definition_id", "_id", "metric_definition");
 
         filters = filters.append(Filters.eq("user_id", userId));
+
+        if(ObjectUtils.allNotNull(start, end) && Utility.isDate(start, end) && Utility.isBefore(start, end)) {
+            filters = filters.append(Filters.and(Filters.gte("time_period_start", Utility.stringToInstant(start)), Filters.lte("time_period_start", Utility.stringToInstant(end))));
+            filters = filters.append(Filters.and(Filters.gte("time_period_end", Utility.stringToInstant(start)), Filters.lte("time_period_end", Utility.stringToInstant(end))));
+        }
 
         Bson regex = Aggregates.match(Filters.and(filters));